requirements.txt
Babel==2.7.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2.5
django-ckeditor==5.7.1
django-cleanup==4.0.0
django-configurations==2.1
django-cors-headers==3.2.0
django-cuser==2017.3.16
django-debug-toolbar==2.0
django-elasticsearch-dsl==7.1.0
django-extensions==2.2.2
django-extra-views==0.12.0
django-filter==2.2.0
django-flatblocks==0.9.4
django-grappelli==2.13.1
django-js-asset==1.2.2
django-modeltranslation==0.13.3
django-mptt==0.10.0
django-multi-form-view==2.0.1
django-multiforloop==0.2.1
django-nested-admin==3.2.4
django-ok-seo==0.8.2
django-phonenumber-field==3.0.1
django-registration==3.0.1
django-reset-migrations==0.4.0
django-rest-framework==0.1.0
django-seo2==1.0.1
django-singleton-model==0.2.3
django-solo==1.1.3
django-tinymce==2.8.0
django-treebeard==4.3
django-unused-media==0.1.13
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
DjangoSEO==1.0
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.17.4
pandas==0.25.3
pbr==5.4.3
phonenumbers==8.10.20
python-dateutil==2.8.1
python-monkey-business==1.0.0
pytils==0.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
stevedore==1.31.0
urllib3==1.25.7
URLObject==2.4.3
validate-email==1.3
virtualenv==16.7.8
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
Werkzeug==0.16.0
whitenoise==4.1.4
